+++ test/jdk/java/lang/ProcessBuilder/InheritIO/InheritIOTest.java
@@ -27,27 +27,28 @@
  * @summary Unit test for java.lang.ProcessBuilder inheritance of standard output and standard error streams
  * @library /test/lib
  * @build jdk.test.lib.process.*
- *        InheritIo
- * @run testng InheritIoTest
+ * @run testng InheritIOTest
  */
 
+import java.util.List;
+import static java.lang.ProcessBuilder.Redirect.INHERIT;
 import jdk.test.lib.process.OutputAnalyzer;
 import jdk.test.lib.process.ProcessTools;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
-import java.util.List;
 import static org.testng.Assert.*;
 
-public class InheritIoTest {
+public class InheritIOTest {
 
+    private static final String EXIT_VALUE_TEMPLATE = "exit value: %d";
     private static final String EXPECTED_RESULT_STDOUT = "message";
-    private static final String EXPECTED_RESULT_STDERR = InheritIo.EXIT_VALUE_TEMPLATE.formatted(0);
+    private static final String EXPECTED_RESULT_STDERR = EXIT_VALUE_TEMPLATE.formatted(0);
 
     @DataProvider
     public Object[][] testCases() {
         return new Object[][]{
-             new Object[] { List.of("InheritIo$TestInheritIo", "printf", EXPECTED_RESULT_STDOUT) },
-             new Object[] { List.of("InheritIo$TestRedirectInherit", "printf", EXPECTED_RESULT_STDOUT) }
+             new Object[] { List.of("InheritIOTest$TestInheritIO", "printf", EXPECTED_RESULT_STDOUT) },
+             new Object[] { List.of("InheritIOTest$TestRedirectInherit", "printf", EXPECTED_RESULT_STDOUT) }
         };
     }
 
@@ -56,9 +57,28 @@ public class InheritIoTest {
         ProcessBuilder processBuilder = ProcessTools.createJavaProcessBuilder(arguments);
         OutputAnalyzer outputAnalyzer = ProcessTools.executeCommand(processBuilder);
         outputAnalyzer.shouldHaveExitValue(0);
-        outputAnalyzer.stderrShouldMatch(EXPECTED_RESULT_STDERR);
-        outputAnalyzer.stdoutShouldMatch(EXPECTED_RESULT_STDOUT);
-        assertEquals(outputAnalyzer.getOutput(),EXPECTED_RESULT_STDOUT + EXPECTED_RESULT_STDERR);
+        assertEquals(outputAnalyzer.getStdout(), EXPECTED_RESULT_STDOUT);
+        assertEquals(outputAnalyzer.getStderr(), EXPECTED_RESULT_STDERR);
+    }
+
+    public static class TestInheritIO {
+        public static void main(String args[]) throws Throwable {
+            int err = new ProcessBuilder(args).inheritIO().start().waitFor();
+            System.err.printf(EXIT_VALUE_TEMPLATE, err);
+            System.exit(err);
+        }
+    }
+
+    public static class TestRedirectInherit {
+        public static void main(String args[]) throws Throwable {
+            int err = new ProcessBuilder(args)
+                    .redirectInput(INHERIT)
+                    .redirectOutput(INHERIT)
+                    .redirectError(INHERIT)
+                    .start().waitFor();
+            System.err.printf(EXIT_VALUE_TEMPLATE, err);
+            System.exit(err);
+        }
     }
 
 }
